backend/requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
pymupdf
spacy
chromadb
python-multipart
httpx
websockets
pydantic
